refactor(cli): extract command helpers and shared arg-parsing utilities from cli.ts
Decompose src/cli.ts (3,600+ lines) by extracting five clusters:

- src/commands/knowledge.ts: readKnowledgeContent, readKnowledgeInput,
  writeMarkdownAsset, normalizeMarkdownAssetName, inferAssetName
- src/commands/remember.ts: resolveRememberContentArg and
  wasRememberFlagValueConsumedAsContent (flag disambiguation guard)
- src/commands/show.ts: normalizeShowArgv (argv preprocessing)
- src/cli/parse-args.ts: hasSubcommand (replaces 5 identical
  hasXxxSubcommand functions) and parsePositiveIntFlag (replaces
  4 identical --limit parsing blocks)

All call sites in cli.ts updated to import from the new locations.
Unused imports (resolveWriteTarget, writeAssetToSource, fetchWebsite-
MarkdownSnapshot, tryReadStdinText, isWithin, MAX_CAPTURED_ASSET_
SLUG_LENGTH) removed.

Fixes #335
